pyproject.toml
[tool.poetry]
name = "crosslink-api"
version = "0.1.0"
description = "crosslink-api"
authors = ["Bin <jaehongbyeon98@gamil.com>"]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.25.0"
fastapi = "^0.105.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
pydantic = "^2.5.3"
sqlalchemy = "^2.0.7"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
pymysql = "^1.1.0"
orjson = "^3.9.10"
requests = "^2.31.0"
pytest = "^7.4.3"
python-multipart = "^0.0.6"
google-auth = "^2.25.2"
python-jose = "^3.3.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
